Admin security short term fix
life_adminlevel variable/function existing on client side is a well
known security hole.  All a script kiddie has to do is use a script
executer to do {life_adminlevel = 1;} and they can do all admin stuff
during that session.  Long term we need to move access to admin
functions to server side and have the server do a RE (remote execution)
of the admin functions on the admin's client.